requirements.txt
PyMuPDF
pytesseract
Pillow
python-docx
